LayoutTests/editing/pasteboard/drag-and-drop-attachment-contenteditable-expected.txt

These tests dragging an attachment element into contenteditable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




PASS target.getElementsByTagName("attachment").length is 1
PASS target.getElementsByTagName("attachment")[0].file.name is "apple.gif"
PASS successfullyParsed is true

TEST COMPLETE
